+++ backend/canisters/group/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Avoid having to specify `authorizer` when pushing notifications ([#7861](https://github.com/open-chat-labs/open-chat/pull/7861))
 - Switch to using LocalUserIndex instead of LocalGroupIndex ([#7917](https://github.com/open-chat-labs/open-chat/pull/7917))
+- Add `requires_captcha` field to prize messages ([#7932](https://github.com/open-chat-labs/open-chat/pull/7932))
 
 ## [[2.0.1727](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1727-group)] - 2025-05-07