+++ backend/canisters/group/CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support bots deleting messages ([#7941](https://github.com/open-chat-labs/open-chat/pull/7941))
 - Support bots adding reactions ([#7944](https://github.com/open-chat-labs/open-chat/pull/7944))
 
+## [[2.0.1738](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1738-group)] - 2025-05-09
+
 ### Changed
 
 - Avoid having to specify `authorizer` when pushing notifications ([#7861](https://github.com/open-chat-labs/open-chat/pull/7861))